@@ -2,8 +2,8 @@ package com.amalbit.animationongooglemap.projectionBased;
 
 import android.os.Build;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.Window;
 import android.view.WindowManager;
 import com.google.android.gms.maps.CameraUpdateFactory;
