@@ -9,8 +9,10 @@ import android.animation.ValueAnimator;
 import android.graphics.DashPathEffect;
 import android.graphics.PathEffect;
 import android.graphics.PathMeasure;
-import android.support.annotation.NonNull;
 import android.view.animation.DecelerateInterpolator;
+
+import androidx.annotation.NonNull;
+
 import com.amalbit.trail.contract.AnimationCallback;
 import com.amalbit.trail.util.AnimatorListener;
 
