@@ -5,213 +5,211 @@ import android.graphics.Canvas;
 import android.graphics.Color;
 import android.graphics.Paint;
 import android.graphics.RectF;
-import android.support.annotation.Nullable;
 import android.util.AttributeSet;
 import android.view.View;
+
+import androidx.annotation.Nullable;
+
 import com.amalbit.trail.contract.GooglemapProvider;
 import com.amalbit.trail.contract.OverlayView;
 import com.google.android.gms.maps.GoogleMap;
+
 import java.util.ArrayList;
 import java.util.List;
 
 public class RouteOverlayView extends View implements OverlayView {
 
-  /**
-   * Debug
-   **/
-  private int debugNumColumns, debugNumRows;
+    private final Object mSvgLock = new Object();
+    /**
+     * Debug
+     **/
+    private int debugNumColumns, debugNumRows;
+    private int debugCellWidth, debugCellHeight;
+    private Paint debugGridPaing = new Paint();
+    private GooglemapProvider googleMapProvider;
+    private List<OverlayPolyline> overlayPolylines;
 
-  private int debugCellWidth, debugCellHeight;
-
-  private Paint debugGridPaing = new Paint();
-
-  private GooglemapProvider googleMapProvider;
-
-  public enum RouteType {
-    PATH,
-    ARC,
-    DASH
-  }
-
-  private final Object mSvgLock = new Object();
-
-  private List<OverlayPolyline> overlayPolylines;
-
-  public RouteOverlayView(Context context, AttributeSet attrs) {
-    super(context, attrs);
-    init(attrs);
-    setUpDebugProps();
-  }
-
-  public RouteOverlayView(Context context) {
-    super(context);
-    init(null);
-    setUpDebugProps();
-  }
-
-  private void init(@Nullable AttributeSet attrSet) {
-    setLayerType(View.LAYER_TYPE_SOFTWARE, null);
-    overlayPolylines = new ArrayList<>();
-  }
-
-  private void setUpDebugProps() {
-    debugGridPaing.setColor(Color.GRAY);
-    debugGridPaing.setStyle(Paint.Style.FILL_AND_STROKE);
-  }
-
-  private void calculateDimensionForDebugGrid() {
-    debugNumColumns = getWidth() / 20;
-    debugNumRows = getHeight() / 20;
-    debugCellWidth = getWidth() / debugNumColumns;
-    debugCellHeight = getHeight() / debugNumRows;
-    invalidate();
-  }
-
-  @Override
-  protected void onSizeChanged(int w, int h, int oldw, int oldh) {
-    super.onSizeChanged(w, h, oldw, oldh);
-    calculateDimensionForDebugGrid();
-  }
-
-  @Override
-  public void addGoogleMapProvider(GooglemapProvider googleMapProvider) {
-    this.googleMapProvider = googleMapProvider;
-  }
-
-  @Override
-  public void onMapReady() {
-
-  }
-
-  @Override
-  public void onCameraMove() {
-    if (isGoogleMapNotNull() && overlayPolylines == null) {
-      return;
-    }
-    GoogleMap googleMap = googleMapProvider.getGoogleMapWeakReference().get();
-    float cameraBearing = googleMap.getCameraPosition().bearing;
-    for (OverlayPolyline overlayPolyline : overlayPolylines) {
-      if (overlayPolyline.getProjectionHelper().getCenterLatLng() != null) {
-        overlayPolyline.getProjectionHelper().onCameraMove(
-            googleMap.getProjection(),
-            googleMap.getCameraPosition(),
-            cameraBearing
-        );
-      }
-      if (overlayPolyline.getShadowProjectionHelper().getCenterLatLng() != null) {
-        overlayPolyline.getShadowProjectionHelper().onCameraMove(
-            googleMap.getProjection(),
-            googleMap.getCameraPosition(),
-            cameraBearing
-        );
-      }
-    }
-  }
-
-  public void removeRoutes() {
-    clearAnimation();
-    stopAllAnimation();
-    overlayPolylines.clear();
-    invalidate();
-  }
-
-  public void removeRoute(OverlayPolyline overlayPolyline) {
-    overlayPolylines.remove(overlayPolyline);
-    invalidate();
-  }
-
-  protected void addPath(OverlayPolyline overlayPolyline) {
-    overlayPolylines.add(overlayPolyline);
-    invalidate();
-    onCameraMove();
-  }
-
-  @Override
-  protected void onDraw(Canvas canvas) {
-    super.onDraw(canvas);
-    synchronized (mSvgLock) {
-      drawPathBorder(canvas);
-      drawRoute(canvas);
-    }
-  }
-
-  private void drawDebugDrid(Canvas canvas) {
-    int width = getWidth();
-    int height = getHeight();
-    for (int i = 1; i < debugNumColumns; i++) {
-      canvas.drawLine(i * debugCellWidth, 0, i * debugCellWidth, height, debugGridPaing);
+    public RouteOverlayView(Context context, AttributeSet attrs) {
+        super(context, attrs);
+        init(attrs);
+//        setUpDebugProps();
     }
 
-    for (int i = 1; i < debugNumRows; i++) {
-      canvas.drawLine(0, i * debugCellHeight, width, i * debugCellHeight, debugGridPaing);
+    public RouteOverlayView(Context context) {
+        super(context);
+        init(null);
+//        setUpDebugProps();
     }
-  }
 
-  private void drawPathBorder(Canvas canvas) {
-    drawDebugDrid(canvas);
-    int count = 0;
-    for (OverlayPolyline overlayPolyline : overlayPolylines) {
-      canvas.drawRect(overlayPolyline.getRectF(), overlayPolyline.getPaintDebug());
-      canvas.drawCircle(overlayPolyline.getRectF().centerX(), overlayPolyline.getRectF().centerY(), 20,
-          overlayPolyline.getTopLayerPaint());
-      canvas.drawText("" + count, overlayPolyline.getRectF().centerX(), overlayPolyline.getRectF().centerY(),
-          overlayPolyline
-              .getPaintDebug());
-      if (overlayPolyline.getProjectionHelper() != null && overlayPolyline.getProjectionHelper().point != null) {
-        canvas.drawCircle(overlayPolyline.getProjectionHelper().point.x, overlayPolyline.getProjectionHelper().point.y,
-            10,
-            overlayPolyline.getBottomLayerPaint());
-        canvas
-            .drawText("" + count, overlayPolyline.getProjectionHelper().point.x,
-                overlayPolyline.getProjectionHelper().point.y, overlayPolyline
-                    .getPaintDebug());
-      }
-      if (overlayPolyline.getShadowDrawPath() != null) {
-        RectF rectF = new RectF();
-        overlayPolyline.getShadowDrawPath().computeBounds(rectF, true);
-        canvas.drawRect(rectF, overlayPolyline.getPaintDebug());
-      }
-      if (overlayPolyline.getShadowProjectionHelper() != null
-          && overlayPolyline.getShadowProjectionHelper().point != null) {
-        canvas.drawCircle(overlayPolyline.getShadowProjectionHelper().point.x,
-            overlayPolyline.getShadowProjectionHelper().point.y, 20,
-            overlayPolyline.getPaintDebug());
-      }
-      count++;
+    private void init(@Nullable AttributeSet attrSet) {
+        setLayerType(View.LAYER_TYPE_SOFTWARE, null);
+        overlayPolylines = new ArrayList<>();
     }
-  }
 
-  private void drawRoute(Canvas canvas) {
-    for (OverlayPolyline overlayPolyline : overlayPolylines) {
-      if (overlayPolyline.getDrawPath() == null) {
-        return;
-      }
-      if (overlayPolyline.getRouteType() == RouteType.ARC) {
-        AnimationArcHelper animationArcHelper = (AnimationArcHelper) overlayPolyline.getAnimationHelper();
-        if (animationArcHelper.animStarted) {
-          if (overlayPolyline.getShadowDrawPath() != null) {
-            canvas.drawPath(overlayPolyline.getShadowDrawPath(),
-                overlayPolyline.getShadowPaint());
-          }
-          if (animationArcHelper.isFirstTimeDrawing) {
-            canvas.drawPath(overlayPolyline.getDrawPath(), overlayPolyline.getTopLayerPaint());
-          } else {
-            canvas.drawPath(overlayPolyline.getDrawPath(), overlayPolyline.getBottomLayerPaint());
-            canvas.drawPath(overlayPolyline.getDrawPath(), overlayPolyline.getTopLayerPaint());
-          }
+    private void setUpDebugProps() {
+        debugGridPaing.setColor(Color.GRAY);
+        debugGridPaing.setStyle(Paint.Style.FILL_AND_STROKE);
+    }
+
+    private void calculateDimensionForDebugGrid() {
+        debugNumColumns = getWidth() / 20;
+        debugNumRows = getHeight() / 20;
+        debugCellWidth = getWidth() / debugNumColumns;
+        debugCellHeight = getHeight() / debugNumRows;
+        invalidate();
+    }
+
+    @Override
+    protected void onSizeChanged(int w, int h, int oldw, int oldh) {
+        super.onSizeChanged(w, h, oldw, oldh);
+        calculateDimensionForDebugGrid();
+    }
+
+    @Override
+    public void addGoogleMapProvider(GooglemapProvider googleMapProvider) {
+        this.googleMapProvider = googleMapProvider;
+    }
+
+    @Override
+    public void onMapReady() {
+
+    }
+
+    @Override
+    public void onCameraMove() {
+        if (isGoogleMapNotNull() && overlayPolylines == null) {
+            return;
         }
-      } else if (overlayPolyline.getRouteType() == RouteType.PATH) {
-        if (((AnimationRouteHelper) overlayPolyline.getAnimationHelper()).isFirstTimeDrawing) {
-          canvas.drawPath(overlayPolyline.getDrawPath(), overlayPolyline.getTopLayerPaint());
-        } else {
-          canvas.drawPath(overlayPolyline.getDrawPath(), overlayPolyline.getBottomLayerPaint());
-          canvas.drawPath(overlayPolyline.getDrawPath(), overlayPolyline.getTopLayerPaint());
+        GoogleMap googleMap = googleMapProvider.getGoogleMapWeakReference().get();
+        float cameraBearing = googleMap.getCameraPosition().bearing;
+        for (OverlayPolyline overlayPolyline : overlayPolylines) {
+            if (overlayPolyline.getProjectionHelper().getCenterLatLng() != null) {
+                overlayPolyline.getProjectionHelper().onCameraMove(
+                        googleMap.getProjection(),
+                        googleMap.getCameraPosition(),
+                        cameraBearing
+                );
+            }
+            if (overlayPolyline.getShadowProjectionHelper().getCenterLatLng() != null) {
+                overlayPolyline.getShadowProjectionHelper().onCameraMove(
+                        googleMap.getProjection(),
+                        googleMap.getCameraPosition(),
+                        cameraBearing
+                );
+            }
         }
-      } else if (overlayPolyline.getRouteType() == RouteType.DASH) {
-        canvas.drawPath(overlayPolyline.getDrawPath(), overlayPolyline.getPaintDash());
-      }
     }
-  }
+
+    public void removeRoutes() {
+        clearAnimation();
+        stopAllAnimation();
+        overlayPolylines.clear();
+        invalidate();
+    }
+
+    public void removeRoute(OverlayPolyline overlayPolyline) {
+        overlayPolylines.remove(overlayPolyline);
+        invalidate();
+    }
+
+    protected void addPath(OverlayPolyline overlayPolyline) {
+        overlayPolylines.add(overlayPolyline);
+        invalidate();
+        onCameraMove();
+    }
+
+    @Override
+    protected void onDraw(Canvas canvas) {
+        super.onDraw(canvas);
+        synchronized (mSvgLock) {
+//            drawPathBorder(canvas);
+            drawRoute(canvas);
+        }
+    }
+
+    private void drawDebugDrid(Canvas canvas) {
+        int width = getWidth();
+        int height = getHeight();
+        for (int i = 1; i < debugNumColumns; i++) {
+            canvas.drawLine(i * debugCellWidth, 0, i * debugCellWidth, height, debugGridPaing);
+        }
+
+        for (int i = 1; i < debugNumRows; i++) {
+            canvas.drawLine(0, i * debugCellHeight, width, i * debugCellHeight, debugGridPaing);
+        }
+    }
+
+    private void drawPathBorder(Canvas canvas) {
+        drawDebugDrid(canvas);
+        int count = 0;
+        for (OverlayPolyline overlayPolyline : overlayPolylines) {
+            canvas.drawRect(overlayPolyline.getRectF(), overlayPolyline.getPaintDebug());
+            canvas.drawCircle(overlayPolyline.getRectF().centerX(), overlayPolyline.getRectF().centerY(), 20,
+                    overlayPolyline.getTopLayerPaint());
+            canvas.drawText("" + count, overlayPolyline.getRectF().centerX(), overlayPolyline.getRectF().centerY(),
+                    overlayPolyline
+                            .getPaintDebug());
+            if (overlayPolyline.getProjectionHelper() != null && overlayPolyline.getProjectionHelper().point != null) {
+                canvas.drawCircle(overlayPolyline.getProjectionHelper().point.x, overlayPolyline.getProjectionHelper().point.y,
+                        10,
+                        overlayPolyline.getBottomLayerPaint());
+                canvas
+                        .drawText("" + count, overlayPolyline.getProjectionHelper().point.x,
+                                overlayPolyline.getProjectionHelper().point.y, overlayPolyline
+                                        .getPaintDebug());
+            }
+            if (overlayPolyline.getShadowDrawPath() != null) {
+                RectF rectF = new RectF();
+                overlayPolyline.getShadowDrawPath().computeBounds(rectF, true);
+                canvas.drawRect(rectF, overlayPolyline.getPaintDebug());
+            }
+            if (overlayPolyline.getShadowProjectionHelper() != null
+                    && overlayPolyline.getShadowProjectionHelper().point != null) {
+                canvas.drawCircle(overlayPolyline.getShadowProjectionHelper().point.x,
+                        overlayPolyline.getShadowProjectionHelper().point.y, 20,
+                        overlayPolyline.getPaintDebug());
+            }
+            count++;
+        }
+    }
+
+    private void drawRoute(Canvas canvas) {
+        for (OverlayPolyline overlayPolyline : overlayPolylines) {
+            if (overlayPolyline.getDrawPath() == null) {
+                return;
+            }
+            if (overlayPolyline.getRouteType() == RouteType.ARC) {
+                AnimationArcHelper animationArcHelper = (AnimationArcHelper) overlayPolyline.getAnimationHelper();
+                if (animationArcHelper.animStarted) {
+                    if (overlayPolyline.getShadowDrawPath() != null) {
+                        canvas.drawPath(overlayPolyline.getShadowDrawPath(),
+                                overlayPolyline.getShadowPaint());
+                    }
+                    if (animationArcHelper.isFirstTimeDrawing) {
+                        canvas.drawPath(overlayPolyline.getDrawPath(), overlayPolyline.getTopLayerPaint());
+                    } else {
+                        canvas.drawPath(overlayPolyline.getDrawPath(), overlayPolyline.getBottomLayerPaint());
+                        canvas.drawPath(overlayPolyline.getDrawPath(), overlayPolyline.getTopLayerPaint());
+                    }
+                }
+            } else if (overlayPolyline.getRouteType() == RouteType.PATH) {
+                if (((AnimationRouteHelper) overlayPolyline.getAnimationHelper()).isFirstTimeDrawing) {
+                    canvas.drawPath(overlayPolyline.getDrawPath(), overlayPolyline.getTopLayerPaint());
+                } else {
+                    canvas.drawPath(overlayPolyline.getDrawPath(), overlayPolyline.getBottomLayerPaint());
+                    canvas.drawPath(overlayPolyline.getDrawPath(), overlayPolyline.getTopLayerPaint());
+                }
+            } else if (overlayPolyline.getRouteType() == RouteType.DASH) {
+                canvas.drawPath(overlayPolyline.getDrawPath(), overlayPolyline.getPaintDash());
+            }
+        }
+    }
+
+    private boolean isGoogleMapNotNull() {
+        return googleMapProvider != null &&
+                googleMapProvider.getGoogleMapWeakReference() != null &&
+                googleMapProvider.getGoogleMapWeakReference().get() != null;
+    }
 
 //  private void drawMarkers(Canvas canvas) {
 //    if (pickUpPoint != null
@@ -234,26 +232,25 @@ public class RouteOverlayView extends View implements OverlayView {
 //    canvas.drawBitmap(bitmap, point.x, point.y, null);
 //  }
 
-  private boolean isGoogleMapNotNull() {
-    return googleMapProvider != null &&
-        googleMapProvider.getGoogleMapWeakReference() != null &&
-        googleMapProvider.getGoogleMapWeakReference().get() != null;
-  }
-
-
-  public void stopAllAnimation() {
-    for (OverlayPolyline overlayPolyline : overlayPolylines) {
-      if (overlayPolyline.getAnimationHelper() != null) {
-        overlayPolyline.getAnimationHelper().stop(() -> {
-        });
-      }
+    public void stopAllAnimation() {
+        for (OverlayPolyline overlayPolyline : overlayPolylines) {
+            if (overlayPolyline.getAnimationHelper() != null) {
+                overlayPolyline.getAnimationHelper().stop(() -> {
+                });
+            }
+        }
     }
-  }
 
-  @Override
-  protected void onDetachedFromWindow() {
-    super.onDetachedFromWindow();
-    stopAllAnimation();
-  }
+    @Override
+    protected void onDetachedFromWindow() {
+        super.onDetachedFromWindow();
+        stopAllAnimation();
+    }
+
+    public enum RouteType {
+        PATH,
+        ARC,
+        DASH
+    }
 
 }
